+++ src/main/java/com/wanted/preonboarding/applicant/ApplicantService.java
@@ -1,8 +1,8 @@
 package com.wanted.preonboarding.applicant;
 
-import com.wanted.preonboarding.recruit.dto.ApplicantApplyingDto;
-import com.wanted.preonboarding.recruit.dto.ApplyRecruitmentResponseDto;
-import com.wanted.preonboarding.recruit.dto.RecruitmentDetailResponseDto;
+import com.wanted.preonboarding.recruitment.dto.ApplicantApplyingDto;
+import com.wanted.preonboarding.recruitment.dto.ApplyRecruitmentResponseDto;
+import com.wanted.preonboarding.recruitment.dto.RecruitmentDetailResponseDto;
 import org.springframework.stereotype.Service;
 
 @Service
