--- src/main/java/com/wanted/preonboarding/recruit/dto/ApplicantApplyingDto.java
@@ -1,4 +1,0 @@
-package com.wanted.preonboarding.recruit.dto;
-
-public class ApplicantApplyingDto {
-}
