+++ src/main/java/com/wanted/preonboarding/recruitment/Recruitment.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit;
+package com.wanted.preonboarding.recruitment;
 
 //채용공고 엔티티
 public class Recruitment {
