+++ src/main/java/com/wanted/preonboarding/recruitment/dto/Applicant.java
@@ -1,0 +1,4 @@
+package com.wanted.preonboarding.recruitment.dto;
+
+public class Applicant {
+}
