+++ src/main/java/com/wanted/preonboarding/recruitment/dto/ApplicantApplyingDto.java
@@ -1,0 +1,4 @@
+package com.wanted.preonboarding.recruitment.dto;
+
+public class ApplicantApplyingDto {
+}
