+++ src/main/java/com/wanted/preonboarding/recruitment/dto/ApplyRecruitmentResponseDto.java
@@ -1,0 +1,4 @@
+package com.wanted.preonboarding.recruitment.dto;
+
+public class ApplyRecruitmentResponseDto {
+}
