+++ src/main/java/com/wanted/preonboarding/recruitment/dto/RecruitmentDetailResponseDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit.dto;
+package com.wanted.preonboarding.recruitment.dto;
 
 //채용공고 상세 데이터 dto
 public class RecruitmentDetailResponseDto {
