+++ src/main/java/com/wanted/preonboarding/recruitment/dto/RecruitmentRegisterDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit.dto;
+package com.wanted.preonboarding.recruitment.dto;
 
 //채용공고 등록 데이터 dto
 public class RecruitmentRegisterDto {
