+++ src/main/java/com/wanted/preonboarding/recruitment/dto/RecruitmentResponseDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit.dto;
+package com.wanted.preonboarding.recruitment.dto;
 
 //채용공고 조회 dto
 //채용공괴 조회 및 검색시 사용
