+++ src/main/java/com/wanted/preonboarding/recruitment/dto/RecruitmentUpdateDto.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.recruit.dto;
+package com.wanted.preonboarding.recruitment.dto;
 
 //채용공고 수정 dto
 public class RecruitmentUpdateDto {
